fix: centralize sqlalchemy test logic
