+++ tests/contrib/sqlalchemy/repository/conftest.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from datetime import datetime
+from typing import TYPE_CHECKING, Any
+from uuid import UUID
 
 import pytest
 
@@ -35,3 +37,53 @@ def _patch_bases(monkeypatch: MonkeyPatch) -> None:
     monkeypatch.setattr(base, "UUIDAuditBase", NewUUIDAuditBase)
     monkeypatch.setattr(base, "BigIntBase", NewBigIntBase)
     monkeypatch.setattr(base, "BigIntAuditBase", NewBigIntAuditBase)
+
+
+@pytest.fixture(name="raw_authors")
+def fx_raw_authors() -> list[dict[str, Any]]:
+    """Unstructured author representations."""
+    return [
+        {
+            "id": UUID("97108ac1-ffcb-411d-8b1e-d9183399f63b"),
+            "name": "Agatha Christie",
+            "dob": "1890-09-15",
+            "created": "0001-01-01T00:00:00",
+            "updated": "0001-01-01T00:00:00",
+        },
+        {
+            "id": "5ef29f3c-3560-4d15-ba6b-a2e5c721e4d2",
+            "name": "Leo Tolstoy",
+            "dob": "1828-09-09",
+            "created": "0001-01-01T00:00:00",
+            "updated": "0001-01-01T00:00:00",
+        },
+    ]
+
+
+@pytest.fixture(name="raw_books")
+def fx_raw_books(raw_authors: list[dict[str, Any]]) -> list[dict[str, Any]]:
+    """Unstructured book representations."""
+    return [
+        {
+            "id": UUID("f34545b9-663c-4fce-915d-dd1ae9cea42a"),
+            "title": "Murder on the Orient Express",
+            "author_id": "97108ac1-ffcb-411d-8b1e-d9183399f63b",
+            "author": raw_authors[0],
+            "created": "0001-01-01T00:00:00",
+            "updated": "0001-01-01T00:00:00",
+        },
+    ]
+
+
+@pytest.fixture(name="raw_log_events")
+def fx_raw_log_events() -> list[dict[str, Any]]:
+    """Unstructured log events representations."""
+    return [
+        {
+            "id": "f34545b9-663c-4fce-915d-dd1ae9cea42a",
+            "logged_at": "0001-01-01T00:00:00",
+            "payload": {"foo": "bar", "baz": datetime.now()},
+            "created": "0001-01-01T00:00:00",
+            "updated": "0001-01-01T00:00:00",
+        },
+    ]
